+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 This is used for recording Invader's changes. This changelog is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+- invader-bitmap: Fixed non-power-of-two bitmaps being marked as power-of-two
+
 ## [0.30.1] - 2020-03-15
 ### Fixed
 - invader-edit-qt: Fixed bitmaps with non-equal height & width being cut off